Rename product.hpp -> prod.hpp
