+++ triqs/mesh/prod.hpp
